feat: #27 SelectSpotLocationFeature 이전
